backend/requirements.txt
django >= 4.1.0, <4.2.0
djangorestframework
django-cors-headers
pyyaml